api-server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.0
notion-client==2.2.1
httpx==0.25.2
python-multipart==0.0.6